update sync script
